Added check if the file exists before trying to rename it.

+++ Rakefile
@@ -5,7 +5,7 @@ def symlink_home(src, dest)
   home_dir = ENV['HOME']
   if( !File.exists?(File.join(home_dir, dest)) || File.symlink?(File.join(home_dir, dest)) )
     # FileUtils.ln_sf was making odd nested links, and this works.
-    FileUtils.rm(File.join(home_dir, dest), :verbose => true) 
+    FileUtils.rm(File.join(home_dir, dest), :verbose => true) if File.exists?(File.join(home_dir, dest))
     FileUtils.ln_s(File.join(File.dirname(__FILE__), src), File.join(home_dir, dest), :verbose => true)
   else
     puts "Unable to symlink #{dest} because it exists and is not a symlink"
